Increase spin box button size
**Notes**
- Increase size of spin box buttons
- Change final time to float

**Testing**
- `./style_check.sh`

**Issues**
- https://github.com/dtsullivan/optgui/issues/41#issue-550585110

+++ Optimization_Interface/include/graphics/view.h
@@ -46,7 +46,7 @@ class View : public QGraphicsView {
     void stepSim();
     void toggleSim();
     void setHorizon(int);
-    void setFinaltime(int);
+    void setFinaltime(double);
     void duplicateSelected();
 
  private:
